+++ src/HearThis/UI/AdministrativeSettings.Designer.cs
@@ -13,9 +13,11 @@ namespace HearThis.UI
 		/// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
 		protected override void Dispose(bool disposing)
 		{
-			if (disposing && (components != null))
+			if (disposing)
 			{
-				components.Dispose();
+				if (components != null)
+					components.Dispose();
+				Program.UnregisterStringsLocalized(HandleStringsLocalized);
 			}
 			base.Dispose(disposing);
 		}
@@ -59,6 +61,11 @@ namespace HearThis.UI
 			this._chkBreakAtQuotes = new System.Windows.Forms.CheckBox();
 			this._chkBreakAtParagraphBreaks = new System.Windows.Forms.CheckBox();
 			this.tabPageInterface = new System.Windows.Forms.TabPage();
+			this._groupAdvancedUI = new System.Windows.Forms.GroupBox();
+			this._tableLayoutPanelAdvancedUI = new System.Windows.Forms.TableLayoutPanel();
+			this._chkEnableClipShifting = new System.Windows.Forms.CheckBox();
+			this._lblShiftClipsMenuWarning = new System.Windows.Forms.Label();
+			this._lblShiftClipsExplanation = new System.Windows.Forms.Label();
 			this._chkShowBookAndChapterLabels = new System.Windows.Forms.CheckBox();
 			this.lblColorSchemeChangeRestartWarning = new System.Windows.Forms.Label();
 			this._cboColorScheme = new System.Windows.Forms.ComboBox();
@@ -77,6 +84,8 @@ namespace HearThis.UI
 			this.tabPagePunctuation.SuspendLayout();
 			this._tableLayoutPanelPunctuation.SuspendLayout();
 			this.tabPageInterface.SuspendLayout();
+			this._groupAdvancedUI.SuspendLayout();
+			this._tableLayoutPanelAdvancedUI.SuspendLayout();
 			((System.ComponentModel.ISupportInitialize)(this.l10NSharpExtender1)).BeginInit();
 			this.SuspendLayout();
 			// 
@@ -529,6 +538,7 @@ namespace HearThis.UI
 			// tabPageInterface
 			// 
 			this.tabPageInterface.BackColor = System.Drawing.SystemColors.ButtonFace;
+			this.tabPageInterface.Controls.Add(this._groupAdvancedUI);
 			this.tabPageInterface.Controls.Add(this._chkShowBookAndChapterLabels);
 			this.tabPageInterface.Controls.Add(this.lblColorSchemeChangeRestartWarning);
 			this.tabPageInterface.Controls.Add(this._cboColorScheme);
@@ -541,6 +551,80 @@ namespace HearThis.UI
 			this.tabPageInterface.Size = new System.Drawing.Size(360, 343);
 			this.tabPageInterface.TabIndex = 3;
 			this.tabPageInterface.Text = "Interface";
+			// 
+			// _groupAdvancedUI
+			// 
+			this._groupAdvancedUI.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
+            | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+			this._groupAdvancedUI.Controls.Add(this._tableLayoutPanelAdvancedUI);
+			this.l10NSharpExtender1.SetLocalizableToolTip(this._groupAdvancedUI, null);
+			this.l10NSharpExtender1.SetLocalizationComment(this._groupAdvancedUI, null);
+			this.l10NSharpExtender1.SetLocalizingId(this._groupAdvancedUI, "AdministrativeSettings._groupAdvancedUI");
+			this._groupAdvancedUI.Location = new System.Drawing.Point(14, 172);
+			this._groupAdvancedUI.Name = "_groupAdvancedUI";
+			this._groupAdvancedUI.Size = new System.Drawing.Size(331, 144);
+			this._groupAdvancedUI.TabIndex = 10;
+			this._groupAdvancedUI.TabStop = false;
+			this._groupAdvancedUI.Text = "Advanced";
+			// 
+			// _tableLayoutPanelAdvancedUI
+			// 
+			this._tableLayoutPanelAdvancedUI.ColumnCount = 1;
+			this._tableLayoutPanelAdvancedUI.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100F));
+			this._tableLayoutPanelAdvancedUI.Controls.Add(this._chkEnableClipShifting, 0, 0);
+			this._tableLayoutPanelAdvancedUI.Controls.Add(this._lblShiftClipsMenuWarning, 0, 2);
+			this._tableLayoutPanelAdvancedUI.Controls.Add(this._lblShiftClipsExplanation, 0, 1);
+			this._tableLayoutPanelAdvancedUI.Dock = System.Windows.Forms.DockStyle.Fill;
+			this._tableLayoutPanelAdvancedUI.Location = new System.Drawing.Point(3, 16);
+			this._tableLayoutPanelAdvancedUI.Name = "_tableLayoutPanelAdvancedUI";
+			this._tableLayoutPanelAdvancedUI.RowCount = 3;
+			this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle());
+			this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle());
+			this._tableLayoutPanelAdvancedUI.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
+			this._tableLayoutPanelAdvancedUI.Size = new System.Drawing.Size(325, 125);
+			this._tableLayoutPanelAdvancedUI.TabIndex = 13;
+			// 
+			// _chkEnableClipShifting
+			// 
+			this._chkEnableClipShifting.AutoSize = true;
+			this.l10NSharpExtender1.SetLocalizableToolTip(this._chkEnableClipShifting, null);
+			this.l10NSharpExtender1.SetLocalizationComment(this._chkEnableClipShifting, null);
+			this.l10NSharpExtender1.SetLocalizingId(this._chkEnableClipShifting, "AdministrativeSettings._chkEnableClipShifting");
+			this._chkEnableClipShifting.Location = new System.Drawing.Point(3, 3);
+			this._chkEnableClipShifting.Name = "_chkEnableClipShifting";
+			this._chkEnableClipShifting.Size = new System.Drawing.Size(125, 17);
+			this._chkEnableClipShifting.TabIndex = 11;
+			this._chkEnableClipShifting.Text = "Enable {0} command";
+			this._chkEnableClipShifting.UseVisualStyleBackColor = true;
+			this._chkEnableClipShifting.CheckedChanged += new System.EventHandler(this.chkEnableClipShifting_CheckedChanged);
+			// 
+			// _lblShiftClipsMenuWarning
+			// 
+			this._lblShiftClipsMenuWarning.AutoSize = true;
+			this._lblShiftClipsMenuWarning.ForeColor = System.Drawing.Color.Red;
+			this.l10NSharpExtender1.SetLocalizableToolTip(this._lblShiftClipsMenuWarning, null);
+			this.l10NSharpExtender1.SetLocalizationComment(this._lblShiftClipsMenuWarning, "Param 0: name of \"Shift Clips\" menu command; Param 1: HearThis (program name)");
+			this.l10NSharpExtender1.SetLocalizingId(this._lblShiftClipsMenuWarning, "AdministrativeSettings._lblShiftClipsMenuWarning");
+			this._lblShiftClipsMenuWarning.Location = new System.Drawing.Point(3, 49);
+			this._lblShiftClipsMenuWarning.Name = "_lblShiftClipsMenuWarning";
+			this._lblShiftClipsMenuWarning.Size = new System.Drawing.Size(310, 65);
+			this._lblShiftClipsMenuWarning.TabIndex = 12;
+			this._lblShiftClipsMenuWarning.Text = resources.GetString("_lblShiftClipsMenuWarning.Text");
+			this._lblShiftClipsMenuWarning.Visible = false;
+			// 
+			// _lblShiftClipsExplanation
+			// 
+			this._lblShiftClipsExplanation.AutoSize = true;
+			this.l10NSharpExtender1.SetLocalizableToolTip(this._lblShiftClipsExplanation, null);
+			this.l10NSharpExtender1.SetLocalizationComment(this._lblShiftClipsExplanation, null);
+			this.l10NSharpExtender1.SetLocalizingId(this._lblShiftClipsExplanation, "AdministrativeSettings._labelShiftClipsExplanation");
+			this._lblShiftClipsExplanation.Location = new System.Drawing.Point(3, 23);
+			this._lblShiftClipsExplanation.Name = "_lblShiftClipsExplanation";
+			this._lblShiftClipsExplanation.Size = new System.Drawing.Size(290, 26);
+			this._lblShiftClipsExplanation.TabIndex = 13;
+			this._lblShiftClipsExplanation.Text = "To use this command, right-click the block slider in the main window.";
+			this._lblShiftClipsExplanation.Visible = false;
 			// 
 			// _chkShowBookAndChapterLabels
 			// 
@@ -614,7 +698,7 @@ namespace HearThis.UI
 			// l10NSharpExtender1
 			// 
 			this.l10NSharpExtender1.LocalizationManagerId = "HearThis";
-			this.l10NSharpExtender1.PrefixForNewItems = "AdministrativeSettings";
+			this.l10NSharpExtender1.PrefixForNewItems = "";
 			// 
 			// AdministrativeSettings
 			// 
@@ -632,7 +716,7 @@ namespace HearThis.UI
 			this.l10NSharpExtender1.SetLocalizingId(this, "RestrictAdministrativeAccess.WindowTitle");
 			this.MaximizeBox = false;
 			this.MinimizeBox = false;
-			this.MinimumSize = new System.Drawing.Size(204, 477);
+			this.MinimumSize = new System.Drawing.Size(234, 477);
 			this.Name = "AdministrativeSettings";
 			this.ShowIcon = false;
 			this.ShowInTaskbar = false;
@@ -651,6 +735,9 @@ namespace HearThis.UI
 			this._tableLayoutPanelPunctuation.PerformLayout();
 			this.tabPageInterface.ResumeLayout(false);
 			this.tabPageInterface.PerformLayout();
+			this._groupAdvancedUI.ResumeLayout(false);
+			this._tableLayoutPanelAdvancedUI.ResumeLayout(false);
+			this._tableLayoutPanelAdvancedUI.PerformLayout();
 			((System.ComponentModel.ISupportInitialize)(this.l10NSharpExtender1)).EndInit();
 			this.ResumeLayout(false);
 
@@ -689,5 +776,10 @@ namespace HearThis.UI
 		private System.Windows.Forms.Label lblInterface;
 		private System.Windows.Forms.Label lblColorSchemeChangeRestartWarning;
 		private System.Windows.Forms.CheckBox _chkShowBookAndChapterLabels;
+		private System.Windows.Forms.GroupBox _groupAdvancedUI;
+		private System.Windows.Forms.Label _lblShiftClipsMenuWarning;
+		private System.Windows.Forms.CheckBox _chkEnableClipShifting;
+		private System.Windows.Forms.TableLayoutPanel _tableLayoutPanelAdvancedUI;
+		private System.Windows.Forms.Label _lblShiftClipsExplanation;
 	}
 }